+++ GooglePay/src/test/java/com/braintreepayments/api/GooglePayActivityResultContractUnitTest.java
@@ -39,7 +39,8 @@ public class GooglePayActivityResultContractUnitTest {
 
         PaymentDataRequest paymentDataRequest = PaymentDataRequest.fromJson(googlePayRequest.toJson());
 
-        GooglePayIntentData intentData = new GooglePayIntentData(1, paymentDataRequest);
+        GooglePayPaymentAuthRequest
+                intentData = new GooglePayPaymentAuthRequest(1, paymentDataRequest);
 
         Context context = ApplicationProvider.getApplicationContext();
 
@@ -58,7 +59,7 @@ public class GooglePayActivityResultContractUnitTest {
 
         GooglePayActivityResultContract sut = new GooglePayActivityResultContract();
 
-        GooglePayResult result = sut.parseResult(RESULT_OK, data);
+        GooglePayPaymentAuthResult result = sut.parseResult(RESULT_OK, data);
         assertNotNull(result.getPaymentData());
         assertNull(result.getError());
     }
@@ -69,7 +70,7 @@ public class GooglePayActivityResultContractUnitTest {
 
         GooglePayActivityResultContract sut = new GooglePayActivityResultContract();
 
-        GooglePayResult result = sut.parseResult(RESULT_CANCELED, data);
+        GooglePayPaymentAuthResult result = sut.parseResult(RESULT_CANCELED, data);
 
         Exception error = result.getError();
         assertTrue(error instanceof UserCanceledException);
@@ -84,7 +85,7 @@ public class GooglePayActivityResultContractUnitTest {
 
         GooglePayActivityResultContract sut = new GooglePayActivityResultContract();
 
-        GooglePayResult result = sut.parseResult(RESULT_ERROR, data);
+        GooglePayPaymentAuthResult result = sut.parseResult(RESULT_ERROR, data);
 
         Exception error = result.getError();
         assertTrue(error instanceof GooglePayException);
@@ -99,7 +100,7 @@ public class GooglePayActivityResultContractUnitTest {
 
         GooglePayActivityResultContract sut = new GooglePayActivityResultContract();
 
-        GooglePayResult result = sut.parseResult(2, data);
+        GooglePayPaymentAuthResult result = sut.parseResult(2, data);
 
         Exception error = result.getError();
         assertTrue(error instanceof BraintreeException);