+++ GooglePay/src/test/java/com/braintreepayments/api/GooglePayClientUnitTest.java
@@ -48,16 +48,16 @@ public class GooglePayClientUnitTest {
     private GooglePayRequest baseRequest;
 
     private GooglePayIsReadyToPayCallback readyToPayCallback;
-    private GooglePayIntentDataCallback intentDataCallback;
-    private GooglePayOnActivityResultCallback activityResultCallback;
+    private GooglePayPaymentAuthRequestCallback intentDataCallback;
+    private GooglePayTokenizeCallback activityResultCallback;
     private ActivityInfo activityInfo;
 
     @Before
     public void beforeEach() {
         activity = mock(FragmentActivity.class);
         readyToPayCallback = mock(GooglePayIsReadyToPayCallback.class);
-        activityResultCallback = mock(GooglePayOnActivityResultCallback.class);
-        intentDataCallback = mock(GooglePayIntentDataCallback.class);
+        activityResultCallback = mock(GooglePayTokenizeCallback.class);
+        intentDataCallback = mock(GooglePayPaymentAuthRequestCallback.class);
         activityInfo = mock(ActivityInfo.class);
 
         baseRequest = new GooglePayRequest();
@@ -184,10 +184,10 @@ public class GooglePayClientUnitTest {
 
     // endregion
 
-    // region requestPayment
+    // region createPaymentAuthRequest
 
     @Test
-    public void requestPayment_callsBackIntentData() throws JSONException {
+    public void createPaymentAuthRequest_callsBackIntentData() throws JSONException {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
                         .environment("sandbox")
@@ -223,13 +223,13 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(googlePayRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(googlePayRequest, intentDataCallback);
 
-        ArgumentCaptor<GooglePayIntentData> captor = ArgumentCaptor.forClass(
-                GooglePayIntentData.class);
-        verify(intentDataCallback).onGooglePayIntentData(captor.capture(), isNull());
+        ArgumentCaptor<GooglePayPaymentAuthRequest> captor = ArgumentCaptor.forClass(
+                GooglePayPaymentAuthRequest.class);
+        verify(intentDataCallback).onResult(captor.capture(), isNull());
 
-        GooglePayIntentData intent = captor.getValue();
+        GooglePayPaymentAuthRequest intent = captor.getValue();
 
         assertEquals(WalletConstants.ENVIRONMENT_TEST, intent.getGooglePayEnvironment());
         PaymentDataRequest paymentDataRequest = intent.getPaymentDataRequest();
@@ -318,7 +318,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_includesATokenizationKeyWhenPresent() throws JSONException {
+    public void createPaymentAuthRequest_includesATokenizationKeyWhenPresent() throws JSONException {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
                         .environment("sandbox")
@@ -345,12 +345,12 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(googlePayRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(googlePayRequest, intentDataCallback);
 
-        ArgumentCaptor<GooglePayIntentData> captor = ArgumentCaptor.forClass(
-                GooglePayIntentData.class);
-        verify(intentDataCallback).onGooglePayIntentData(captor.capture(), isNull());
-        GooglePayIntentData intent = captor.getValue();
+        ArgumentCaptor<GooglePayPaymentAuthRequest> captor = ArgumentCaptor.forClass(
+                GooglePayPaymentAuthRequest.class);
+        verify(intentDataCallback).onResult(captor.capture(), isNull());
+        GooglePayPaymentAuthRequest intent = captor.getValue();
 
         PaymentDataRequest paymentDataRequest = intent.getPaymentDataRequest();
         JSONObject paymentDataRequestJson = new JSONObject(paymentDataRequest.toJson());
@@ -366,7 +366,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_doesNotIncludeATokenizationKeyWhenNotPresent() throws JSONException {
+    public void createPaymentAuthRequest_doesNotIncludeATokenizationKeyWhenNotPresent() throws JSONException {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
                         .environment("sandbox")
@@ -393,12 +393,12 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(googlePayRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(googlePayRequest, intentDataCallback);
 
-        ArgumentCaptor<GooglePayIntentData> captor = ArgumentCaptor.forClass(
-                GooglePayIntentData.class);
-        verify(intentDataCallback).onGooglePayIntentData(captor.capture(), isNull());
-        GooglePayIntentData intent = captor.getValue();
+        ArgumentCaptor<GooglePayPaymentAuthRequest> captor = ArgumentCaptor.forClass(
+                GooglePayPaymentAuthRequest.class);
+        verify(intentDataCallback).onResult(captor.capture(), isNull());
+        GooglePayPaymentAuthRequest intent = captor.getValue();
 
         PaymentDataRequest paymentDataRequest = intent.getPaymentDataRequest();
         JSONObject paymentDataRequestJson = new JSONObject(paymentDataRequest.toJson());
@@ -413,7 +413,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_sendsAnalyticsEvent() {
+    public void createPaymentAuthRequest_sendsAnalyticsEvent() {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
                         .environment("sandbox")
@@ -441,7 +441,7 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(googlePayRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(googlePayRequest, intentDataCallback);
 
         InOrder order = inOrder(braintreeClient);
         order.verify(braintreeClient).sendAnalyticsEvent(eq("google-payment.selected"));
@@ -449,7 +449,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_postsExceptionWhenTransactionInfoIsNull() {
+    public void createPaymentAuthRequest_postsExceptionWhenTransactionInfoIsNull() {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
                         .environment("sandbox")
@@ -471,7 +471,7 @@ public class GooglePayClientUnitTest {
         GooglePayRequest googlePayRequest = new GooglePayRequest();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(googlePayRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(googlePayRequest, intentDataCallback);
 
         InOrder order = inOrder(braintreeClient);
         order.verify(braintreeClient).sendAnalyticsEvent(eq("google-payment.selected"));
@@ -479,7 +479,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_whenMerchantNotConfigured_returnsExceptionToFragment() {
+    public void createPaymentAuthRequest_whenMerchantNotConfigured_returnsExceptionToFragment() {
         Configuration configuration = new TestConfigurationBuilder().buildConfiguration();
 
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
@@ -492,11 +492,11 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(baseRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(baseRequest, intentDataCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(
                 Exception.class);
-        verify(intentDataCallback).onGooglePayIntentData(isNull(), captor.capture());
+        verify(intentDataCallback).onResult(isNull(), captor.capture());
 
         assertTrue(captor.getValue() instanceof BraintreeException);
         assertEquals(
@@ -505,7 +505,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_whenSandbox_setsTestEnvironment() throws JSONException {
+    public void createPaymentAuthRequest_whenSandbox_setsTestEnvironment() throws JSONException {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
                         .environment("sandbox")
@@ -526,12 +526,12 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(baseRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(baseRequest, intentDataCallback);
 
-        ArgumentCaptor<GooglePayIntentData> captor = ArgumentCaptor.forClass(
-                GooglePayIntentData.class);
-        verify(intentDataCallback).onGooglePayIntentData(captor.capture(), isNull());
-        GooglePayIntentData intent = captor.getValue();
+        ArgumentCaptor<GooglePayPaymentAuthRequest> captor = ArgumentCaptor.forClass(
+                GooglePayPaymentAuthRequest.class);
+        verify(intentDataCallback).onResult(captor.capture(), isNull());
+        GooglePayPaymentAuthRequest intent = captor.getValue();
 
         PaymentDataRequest paymentDataRequest = intent.getPaymentDataRequest();
         JSONObject paymentDataRequestJson = new JSONObject(paymentDataRequest.toJson());
@@ -541,7 +541,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_whenProduction_setsProductionEnvironment() throws JSONException {
+    public void createPaymentAuthRequest_whenProduction_setsProductionEnvironment() throws JSONException {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
                         .environment("production")
@@ -562,12 +562,12 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(baseRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(baseRequest, intentDataCallback);
 
-        ArgumentCaptor<GooglePayIntentData> captor = ArgumentCaptor.forClass(
-                GooglePayIntentData.class);
-        verify(intentDataCallback).onGooglePayIntentData(captor.capture(), isNull());
-        GooglePayIntentData intent = captor.getValue();
+        ArgumentCaptor<GooglePayPaymentAuthRequest> captor = ArgumentCaptor.forClass(
+                GooglePayPaymentAuthRequest.class);
+        verify(intentDataCallback).onResult(captor.capture(), isNull());
+        GooglePayPaymentAuthRequest intent = captor.getValue();
 
         PaymentDataRequest paymentDataRequest = intent.getPaymentDataRequest();
         JSONObject paymentDataRequestJson = new JSONObject(paymentDataRequest.toJson());
@@ -577,7 +577,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_withGoogleMerchantId_sendGoogleMerchantId() throws JSONException {
+    public void createPaymentAuthRequest_withGoogleMerchantId_sendGoogleMerchantId() throws JSONException {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
                         .environment("sandbox")
@@ -600,12 +600,12 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(baseRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(baseRequest, intentDataCallback);
 
-        ArgumentCaptor<GooglePayIntentData> captor = ArgumentCaptor.forClass(
-                GooglePayIntentData.class);
-        verify(intentDataCallback).onGooglePayIntentData(captor.capture(), isNull());
-        GooglePayIntentData intent = captor.getValue();
+        ArgumentCaptor<GooglePayPaymentAuthRequest> captor = ArgumentCaptor.forClass(
+                GooglePayPaymentAuthRequest.class);
+        verify(intentDataCallback).onResult(captor.capture(), isNull());
+        GooglePayPaymentAuthRequest intent = captor.getValue();
 
         PaymentDataRequest paymentDataRequest = intent.getPaymentDataRequest();
         JSONObject paymentDataRequestJson = new JSONObject(paymentDataRequest.toJson());
@@ -616,7 +616,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_withGoogleMerchantName_sendGoogleMerchantName()
+    public void createPaymentAuthRequest_withGoogleMerchantName_sendGoogleMerchantName()
             throws JSONException {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
@@ -640,12 +640,12 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(baseRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(baseRequest, intentDataCallback);
 
-        ArgumentCaptor<GooglePayIntentData> captor = ArgumentCaptor.forClass(
-                GooglePayIntentData.class);
-        verify(intentDataCallback).onGooglePayIntentData(captor.capture(), isNull());
-        GooglePayIntentData intent = captor.getValue();
+        ArgumentCaptor<GooglePayPaymentAuthRequest> captor = ArgumentCaptor.forClass(
+                GooglePayPaymentAuthRequest.class);
+        verify(intentDataCallback).onResult(captor.capture(), isNull());
+        GooglePayPaymentAuthRequest intent = captor.getValue();
 
         PaymentDataRequest paymentDataRequest = intent.getPaymentDataRequest();
         JSONObject paymentDataRequestJson = new JSONObject(paymentDataRequest.toJson());
@@ -656,7 +656,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_whenGooglePayCanProcessPayPal_tokenizationPropertiesIncludePayPal()
+    public void createPaymentAuthRequest_whenGooglePayCanProcessPayPal_tokenizationPropertiesIncludePayPal()
             throws JSONException {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
@@ -681,12 +681,12 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(baseRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(baseRequest, intentDataCallback);
 
-        ArgumentCaptor<GooglePayIntentData> captor = ArgumentCaptor.forClass(
-                GooglePayIntentData.class);
-        verify(intentDataCallback).onGooglePayIntentData(captor.capture(), isNull());
-        GooglePayIntentData intent = captor.getValue();
+        ArgumentCaptor<GooglePayPaymentAuthRequest> captor = ArgumentCaptor.forClass(
+                GooglePayPaymentAuthRequest.class);
+        verify(intentDataCallback).onResult(captor.capture(), isNull());
+        GooglePayPaymentAuthRequest intent = captor.getValue();
 
         PaymentDataRequest paymentDataRequest = intent.getPaymentDataRequest();
         JSONObject paymentDataRequestJson = new JSONObject(paymentDataRequest.toJson());
@@ -701,7 +701,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_whenPayPalDisabledByRequest_tokenizationPropertiesLackPayPal()
+    public void createPaymentAuthRequest_whenPayPalDisabledByRequest_tokenizationPropertiesLackPayPal()
             throws JSONException {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
@@ -727,12 +727,12 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(baseRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(baseRequest, intentDataCallback);
 
-        ArgumentCaptor<GooglePayIntentData> captor = ArgumentCaptor.forClass(
-                GooglePayIntentData.class);
-        verify(intentDataCallback).onGooglePayIntentData(captor.capture(), isNull());
-        GooglePayIntentData intent = captor.getValue();
+        ArgumentCaptor<GooglePayPaymentAuthRequest> captor = ArgumentCaptor.forClass(
+                GooglePayPaymentAuthRequest.class);
+        verify(intentDataCallback).onResult(captor.capture(), isNull());
+        GooglePayPaymentAuthRequest intent = captor.getValue();
 
         PaymentDataRequest paymentDataRequest = intent.getPaymentDataRequest();
         JSONObject paymentDataRequestJson = new JSONObject(paymentDataRequest.toJson());
@@ -746,7 +746,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_whenPayPalDisabledInConfigurationAndGooglePayHasPayPalClientId_tokenizationPropertiesContainPayPal()
+    public void createPaymentAuthRequest_whenPayPalDisabledInConfigurationAndGooglePayHasPayPalClientId_tokenizationPropertiesContainPayPal()
             throws JSONException {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
@@ -773,12 +773,12 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(baseRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(baseRequest, intentDataCallback);
 
-        ArgumentCaptor<GooglePayIntentData> captor = ArgumentCaptor.forClass(
-                GooglePayIntentData.class);
-        verify(intentDataCallback).onGooglePayIntentData(captor.capture(), isNull());
-        GooglePayIntentData intent = captor.getValue();
+        ArgumentCaptor<GooglePayPaymentAuthRequest> captor = ArgumentCaptor.forClass(
+                GooglePayPaymentAuthRequest.class);
+        verify(intentDataCallback).onResult(captor.capture(), isNull());
+        GooglePayPaymentAuthRequest intent = captor.getValue();
 
         PaymentDataRequest paymentDataRequest = intent.getPaymentDataRequest();
         JSONObject paymentDataRequestJson = new JSONObject(paymentDataRequest.toJson());
@@ -794,7 +794,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_usesGooglePayConfigurationClientId() throws JSONException {
+    public void createPaymentAuthRequest_usesGooglePayConfigurationClientId() throws JSONException {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
                         .environment("sandbox")
@@ -820,12 +820,12 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(baseRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(baseRequest, intentDataCallback);
 
-        ArgumentCaptor<GooglePayIntentData> captor = ArgumentCaptor.forClass(
-                GooglePayIntentData.class);
-        verify(intentDataCallback).onGooglePayIntentData(captor.capture(), isNull());
-        GooglePayIntentData intent = captor.getValue();
+        ArgumentCaptor<GooglePayPaymentAuthRequest> captor = ArgumentCaptor.forClass(
+                GooglePayPaymentAuthRequest.class);
+        verify(intentDataCallback).onResult(captor.capture(), isNull());
+        GooglePayPaymentAuthRequest intent = captor.getValue();
 
         PaymentDataRequest paymentDataRequest = intent.getPaymentDataRequest();
         JSONObject paymentDataRequestJson = new JSONObject(paymentDataRequest.toJson());
@@ -850,7 +850,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_whenGooglePayConfigurationLacksClientId_tokenizationPropertiesLackPayPal()
+    public void createPaymentAuthRequest_whenGooglePayConfigurationLacksClientId_tokenizationPropertiesLackPayPal()
             throws JSONException {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
@@ -874,12 +874,12 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(baseRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(baseRequest, intentDataCallback);
 
-        ArgumentCaptor<GooglePayIntentData> captor = ArgumentCaptor.forClass(
-                GooglePayIntentData.class);
-        verify(intentDataCallback).onGooglePayIntentData(captor.capture(), isNull());
-        GooglePayIntentData intent = captor.getValue();
+        ArgumentCaptor<GooglePayPaymentAuthRequest> captor = ArgumentCaptor.forClass(
+                GooglePayPaymentAuthRequest.class);
+        verify(intentDataCallback).onResult(captor.capture(), isNull());
+        GooglePayPaymentAuthRequest intent = captor.getValue();
 
         PaymentDataRequest paymentDataRequest = intent.getPaymentDataRequest();
         JSONObject paymentDataRequestJson = new JSONObject(paymentDataRequest.toJson());
@@ -896,7 +896,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_whenConfigurationContainsElo_addsEloAndEloDebitToAllowedPaymentMethods()
+    public void createPaymentAuthRequest_whenConfigurationContainsElo_addsEloAndEloDebitToAllowedPaymentMethods()
             throws JSONException {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
@@ -920,12 +920,12 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(baseRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(baseRequest, intentDataCallback);
 
-        ArgumentCaptor<GooglePayIntentData> captor = ArgumentCaptor.forClass(
-                GooglePayIntentData.class);
-        verify(intentDataCallback).onGooglePayIntentData(captor.capture(), isNull());
-        GooglePayIntentData intent = captor.getValue();
+        ArgumentCaptor<GooglePayPaymentAuthRequest> captor = ArgumentCaptor.forClass(
+                GooglePayPaymentAuthRequest.class);
+        verify(intentDataCallback).onResult(captor.capture(), isNull());
+        GooglePayPaymentAuthRequest intent = captor.getValue();
 
         PaymentDataRequest paymentDataRequest = intent.getPaymentDataRequest();
         JSONObject paymentDataRequestJson = new JSONObject(paymentDataRequest.toJson());
@@ -942,7 +942,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_whenRequestIsNull_forwardsExceptionToListener() {
+    public void createPaymentAuthRequest_whenRequestIsNull_forwardsExceptionToListener() {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
                         .environment("sandbox")
@@ -963,11 +963,11 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(null, intentDataCallback);
+        sut.createPaymentAuthRequest(null, intentDataCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(
                 Exception.class);
-        verify(intentDataCallback).onGooglePayIntentData(isNull(), captor.capture());
+        verify(intentDataCallback).onResult(isNull(), captor.capture());
 
         Exception exception = captor.getValue();
         assertTrue(exception instanceof BraintreeException);
@@ -975,7 +975,7 @@ public class GooglePayClientUnitTest {
     }
 
     @Test
-    public void requestPayment_whenManifestInvalid_forwardsExceptionToListener() {
+    public void createPaymentAuthRequest_whenManifestInvalid_forwardsExceptionToListener() {
         Configuration configuration = new TestConfigurationBuilder()
                 .googlePay(new TestConfigurationBuilder.TestGooglePayConfigurationBuilder()
                         .environment("sandbox")
@@ -996,11 +996,11 @@ public class GooglePayClientUnitTest {
                 new MockGooglePayInternalClientBuilder().build();
 
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
-        sut.requestPayment(baseRequest, intentDataCallback);
+        sut.createPaymentAuthRequest(baseRequest, intentDataCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(
                 Exception.class);
-        verify(intentDataCallback).onGooglePayIntentData(isNull(), captor.capture());
+        verify(intentDataCallback).onResult(isNull(), captor.capture());
 
         Exception exception = captor.getValue();
         assertTrue(exception instanceof BraintreeException);
@@ -1098,9 +1098,9 @@ public class GooglePayClientUnitTest {
 
         String paymentDataJson = Fixtures.RESPONSE_GOOGLE_PAY_CARD;
         PaymentData paymentData = PaymentData.fromJson(paymentDataJson);
-        GooglePayResult
-                googlePayResult = new GooglePayResult(paymentData, null);
-        sut.tokenize(googlePayResult, activityResultCallback);
+        GooglePayPaymentAuthResult
+                googlePayPaymentAuthResult = new GooglePayPaymentAuthResult(paymentData, null);
+        sut.tokenize(googlePayPaymentAuthResult, activityResultCallback);
 
         verify(braintreeClient).sendAnalyticsEvent(eq("google-payment.authorized"));
         ArgumentCaptor<PaymentMethodNonce> captor =
@@ -1122,9 +1122,9 @@ public class GooglePayClientUnitTest {
         GooglePayClient sut = new GooglePayClient(braintreeClient, internalGooglePayClient);
 
         Exception error = new Exception("Error");
-        GooglePayResult
-                googlePayResult = new GooglePayResult(null, error);
-        sut.tokenize(googlePayResult, activityResultCallback);
+        GooglePayPaymentAuthResult
+                googlePayPaymentAuthResult = new GooglePayPaymentAuthResult(null, error);
+        sut.tokenize(googlePayPaymentAuthResult, activityResultCallback);
 
         verify(braintreeClient).sendAnalyticsEvent(eq("google-payment.failed"));
         verify(activityResultCallback).onResult(null, error);
@@ -1140,9 +1140,9 @@ public class GooglePayClientUnitTest {
 
         UserCanceledException userCanceledError =
                 new UserCanceledException("User canceled Google Pay.");
-        GooglePayResult
-                googlePayResult = new GooglePayResult(null, userCanceledError);
-        sut.tokenize(googlePayResult, activityResultCallback);
+        GooglePayPaymentAuthResult
+                googlePayPaymentAuthResult = new GooglePayPaymentAuthResult(null, userCanceledError);
+        sut.tokenize(googlePayPaymentAuthResult, activityResultCallback);
 
         verify(braintreeClient).sendAnalyticsEvent(eq("google-payment.canceled"));
         verify(activityResultCallback).onResult(null, userCanceledError);